Tried to combine folder restructuring
