+++ src/app/shared/components/navigation/navigation.component.html
@@ -1,10 +1,5 @@
 <nav class="navbar">
   <div class="navbar">
-    <div class="corporate-logo" [class.active]="isSidebarVisible">
-      <img src="../../../assets/png/logo.png" height="64px">
-      @if(isSidebarVisible){
-      <img src="../../../assets/png/company-title.png" height="64px">}
-    </div>
     <div class="nav-links">
       <div class="menu-container" [class.active]="isSidebarVisible">
         <button class="menu" (click)="toggleSidebar()">☰</button>
@@ -44,32 +39,4 @@
     }}
   }
 
-  <!-- <a routerLink="Production" routerLinkActive="active" (click)="toggleSidebar()">
-    <img src="../../../assets/svg/production.svg" alt="" >
-    @if(isSidebarVisible){<span class="sidebar-text">Production</span>}
-  </a>
-  <a routerLink="quality" routerLinkActive="active" (click)="toggleSidebar()">
-    <img src="../../../assets/svg/star.svg" alt="" >
-    @if(isSidebarVisible){<span class="sidebar-text">Quality</span>}
-  </a>
-  <a routerLink="" routerLinkActive="active" (click)="toggleSidebar()">
-    <img src="../../../assets/svg/list-square.svg" alt="" >
-    @if(isSidebarVisible){<span class="sidebar-text">Inventory</span>}
-  </a>
-  <a routerLink="quality" routerLinkActive="active" (click)="toggleSidebar()">
-    <img src="../../../assets/svg/piggy.svg" alt="" >
-    @if(isSidebarVisible){<span class="sidebar-text">Material Cost</span>}
-  </a>
-  <a routerLink="quality" routerLinkActive="active" (click)="toggleSidebar()">
-    <img src="../../../assets/svg/dollar.svg" alt="" >
-    @if(isSidebarVisible){<span class="sidebar-text">Sales</span>}
-  </a>
-  <a routerLink="quality" routerLinkActive="active" (click)="toggleSidebar()">
-    <img src="../../../assets/svg/phone.svg" alt="" >
-    @if(isSidebarVisible){<span class="sidebar-text">Contacts</span>}
-  </a>
-  <a routerLink="quality" routerLinkActive="active" (click)="toggleSidebar()">
-    <img src="../../../assets/svg/layers.svg" alt="" >
-    @if(isSidebarVisible){<span class="sidebar-text">Etc</span>}
-  </a> -->
 </div>
