+++ src/app/shared/components/navigation/navigation.component.scss
@@ -25,6 +25,7 @@
 .corporate-logo.active {
   width: 11rem;
   transition: width 0.3s ease;
+}
 .navcontainer.active{
   margin: 0 0 0 13rem
 }
