+++ src/app/shared/components/navigation/navigation.component.ts
@@ -3,6 +3,7 @@ import { DepartmentService } from '../../../core/services/department.service';
 import { Subscription } from 'rxjs';
 import { Department } from '../../models/department.model';
 import { Chart } from '../../models/chart.model';
+import { SidebarService } from '../../../core/services/sidebar.service';
 
 @Component({
   selector: 'app-navigation',
@@ -17,7 +18,10 @@ export class NavigationComponent implements OnInit, OnDestroy{
   menuDepartments:Department[] | null = [];
   private userDepartmentSubscription!: Subscription;
 
-  constructor(private departmentService:DepartmentService) {
+  constructor(
+    private departmentService:DepartmentService,
+    private sidebar:SidebarService
+  ) {
   }
 
   ngOnInit(): void {
@@ -42,17 +46,13 @@ export class NavigationComponent implements OnInit, OnDestroy{
     }
   }
 
+
+  checkSidebar(): boolean{
+    return this.sidebar.isSidebarVisible
+  }
+
   toggleSidebar(){
-    let mySVG = document.getElementById('mySVG')
-    console.log(mySVG?.getAttribute("margin-right"))
-    this.isSidebarVisible = !this.isSidebarVisible
-    if(this.isSidebarVisible){
-      mySVG?.setAttribute("width", "-150")
-
-    } else {
-      mySVG?.setAttribute("viewBox", "0")
-    };
-
+    this.sidebar.toggleSidebar()
     setTimeout(() => {
       const sidebarTexts = document.querySelectorAll('.sidebar-text');
       sidebarTexts.forEach((element: Element) => {
@@ -61,4 +61,5 @@ export class NavigationComponent implements OnInit, OnDestroy{
       });
     }, 100);
   }
+
 }
