+++ src/app/shared/components/sidebar/sidebar.component.scss
