--- src/app/shared/navigation/navigation.component.html
@@ -1,29 +1,0 @@
-<nav class="navbar">
-  <div class="navbar">
-    <div class="corporate-logo" [class.active]="checkSidebar()">
-      <img src="../../../assets/png/logo.png" height="64px">
-      @if(checkSidebar()){
-      <img src="../../../assets/png/company-title.png" height="64px">}
-    </div>
-    <div class="nav-links">
-      <div class="menu-container" [class.active]="checkSidebar()">
-        <button class="menu" (click)="toggleSidebar()">☰</button>
-      </div>
-    </div>
-  </div>
-  <div class="top-right">
-
-      <a class="search">
-        <img src="../../../assets/svg/search.svg" alt="search">
-      </a>
-
-      <a class="notification-bell">
-        <img src="../../../assets/svg/notification-bell.svg" alt="notification">
-      </a>
-
-      <button class="pfp" routerLink="profile">
-        <img class="profile-pic" src="../../../assets/pfp/download.jpg" alt="pfp" width="50px">
-      </button>
-
-  </div>
-</nav>
